nebula_status: Refactor Status for arbitrary data
The Status type now accepts arbitrary data, so long as it can be
converted into the Bytes type from the bytes crate. It defaults to
holding Strings, but can now hold any data type.

When the Status is created, a copy of the provided data as Bytes is
stored into a separate field to prevent unnecessary clones by consumer
programs.

Documentation was also updated and created to reflect these changes, and
warnings were fixed.