nebula_rpc: add crate based on tonic
Created RPC bindings and related types. Rust programs that implement a
handler server can use this crate to ease development.

Most testing is going to need to be as integration tests, and those are
not written yet. This code compiles, but has not yet been checked, since
the focus was on exploring the tonic library and seeing if a useful
crate could be created.

Helps toward #5 and #1.