validate-handler: refactor field validators
The existing implementation of `Validator` for the field types was not
object-safe, and thus would not compile as written.

This revises the implementation to make it object-safe, gets the code
compiling with passing tests, and fixes most warnings that pop up. Some
warnings are ignored due to the in-development nature of this crate.